Debugging code for consistency in bufferevent callback arg types
There are a ton of cases where we cast something to void* for a
callback, and get that same thing back later.  It shouldn't be a
problem with correct C struct layout, but it's a potential trouble
source for the future.  This patch adds deliberate misalignment for
the different bufferevent structs, so we can't accidentally type-pun
them in callbacks, and adds checks to make sure we're getting the
right types.